FIAP/Enterprise Challenge - Sprint 1/python/requirements.txt
paho-mqtt
boto3
pandas
pyarrow
influxdb
psycopg2-binary
APScheduler
scikit-learn
joblib
requests
python-dotenv
terraform-compliance